style: 티커별 뉴스 요약 컴포넌트 A/B 테스트를 위한 B안 UI 구현
