+++ src/components/Detail/KeywordBubbleMap/index.tsx
@@ -32,6 +32,8 @@ import {
   pillWidth,
   truncateTitle,
 } from './positions';
+import { Paragraph } from '@/components/common/typography/Paragraph';
+import useIsMobile from '@/hooks/useIsMobile';
 
 type Props = {
   positiveRatio: number;
@@ -50,6 +52,7 @@ export default function KeywordBubbleMap({
   onNewsClick,
   date,
 }: Props) {
+  const isMobile = useIsMobile();
   const uid = useRef(Math.random().toString(36).slice(2, 8)).current;
   const [selected, setSelected] = useState<KeywordsWithArticleResponse | null>(
     null
@@ -85,308 +88,332 @@ export default function KeywordBubbleMap({
   const negShadowId = `${uid}-sn`;
   const newsShadowId = `${uid}-ns`;
 
+  const formatDate = (dateString: string) => {
+    const [, month, day] = dateString.split('-');
+    return `${Number(month)}월 ${Number(day)}일`;
+  };
+
   return (
-    <Container onClick={close}>
-      <Background $ratio={positiveRatio} />
-      <DateText>{date}</DateText>
+    <>
+      <Paragraph size={isMobile ? 'xs' : 's'} weight="bold">
+        {formatDate(date)}의 뉴스 요약
+      </Paragraph>
+      <Container onClick={close}>
+        <Background $ratio={positiveRatio} />
+        <DateText>{date}</DateText>
 
-      <Svg
-        viewBox={`0 0 ${SVG_WIDTH} ${SVG_HEIGHT}`}
-        preserveAspectRatio="xMidYMid meet"
-        style={{ fontFamily: 'Pretendard, sans-serif' }}
-      >
-        <defs>
-          <filter id={posShadowId} x="-30%" y="-30%" width="160%" height="160%">
-            <feDropShadow
-              dx="2"
-              dy="2"
-              stdDeviation="2.5"
-              floodColor="rgba(237,202,206,0.35)"
-            />
-          </filter>
-          <filter id={negShadowId} x="-30%" y="-30%" width="160%" height="160%">
-            <feDropShadow
-              dx="2"
-              dy="2"
-              stdDeviation="2.5"
-              floodColor="rgba(169,204,253,0.35)"
-            />
-          </filter>
-
-          <filter
-            id={newsShadowId}
-            x="-20%"
-            y="-40%"
-            width="140%"
-            height="180%"
-          >
-            <feDropShadow
-              dx="0"
-              dy="3"
-              stdDeviation="5"
-              floodColor="rgba(0,0,0,0.10)"
-            />
-          </filter>
-        </defs>
-
-        <AnimatePresence>
-          {selected !== null && (
-            <motion.rect
-              key="dim"
-              x={0}
-              y={0}
-              width={SVG_WIDTH}
-              height={SVG_HEIGHT}
-              initial={{ opacity: 0 }}
-              animate={{ opacity: 1 }}
-              exit={{ opacity: 0 }}
-              transition={{ duration: 0.3 }}
-              fill={isPos ? 'rgba(255,245,247,0.94)' : 'rgba(245,250,255,0.94)'}
-              style={{ cursor: 'default' }}
-              onClick={(e) => {
-                e.stopPropagation();
-                close();
-              }}
-            />
-          )}
-        </AnimatePresence>
-
-        {positiveKeywords.map((kw, i) => {
-          const { x, y } = posPos[i];
-          const pw = pillWidth(kw.keyword);
-          const isSelected = selected?.keyword === kw.keyword;
-          const isDimmed = selected !== null && !isSelected;
-          const hasArticles = kw.articles.length > 0;
-          const fill = hasArticles ? POS_FILL : EMPTY_FILL;
-          const stroke = hasArticles ? POS_STROKE : EMPTY_STROKE;
-          const textFill = hasArticles ? POS_TEXT : EMPTY_TEXT;
-
-          return (
-            <motion.g
-              key={kw.keyword}
-              role="button"
-              tabIndex={isDimmed ? -1 : 0}
-              animate={{
-                x: isSelected ? SVG_WIDTH / 2 - x : 0,
-                y: isSelected ? SVG_CENTER_Y - y : 0,
-                opacity: isDimmed ? 0 : 1,
-              }}
-              transition={{ duration: 0.35, ease: 'easeOut' }}
-              whileHover={
-                selected === null && hasArticles ? { scale: 1.08 } : {}
-              }
-              whileFocus={
-                selected === null && hasArticles ? { scale: 1.08 } : {}
-              }
-              style={{
-                transformOrigin: `${x}px ${y}px`,
-                cursor:
-                  selected === null && hasArticles ? 'pointer' : 'default',
-                outline: 'none',
-              }}
-              onClick={(e) => {
-                if (selected === null && hasArticles) {
-                  e.stopPropagation();
-                  setSelected(kw);
-                }
-              }}
-              onKeyDown={(e) => {
-                if (
-                  selected === null &&
-                  hasArticles &&
-                  (e.key === 'Enter' || e.key === ' ')
-                ) {
-                  e.preventDefault();
-                  e.stopPropagation();
-                  setSelected(kw);
-                }
-              }}
+        <Svg
+          viewBox={`0 0 ${SVG_WIDTH} ${SVG_HEIGHT}`}
+          preserveAspectRatio="xMidYMid meet"
+          style={{ fontFamily: 'Pretendard, sans-serif' }}
+        >
+          <defs>
+            <filter
+              id={posShadowId}
+              x="-30%"
+              y="-30%"
+              width="160%"
+              height="160%"
             >
-              <rect
-                x={x - pw / 2}
-                y={y - PILL_HEIGHT / 2}
-                width={pw}
-                height={PILL_HEIGHT}
-                rx={PILL_RADIUS}
-                fill={fill}
-                stroke={stroke}
-                strokeWidth={1}
-                filter={`url(#${posShadowId})`}
+              <feDropShadow
+                dx="2"
+                dy="2"
+                stdDeviation="2.5"
+                floodColor="rgba(237,202,206,0.35)"
               />
-              <text
-                x={x}
-                y={y}
-                textAnchor="middle"
-                dominantBaseline="middle"
-                fontSize={13}
-                fontWeight="600"
-                fill={textFill}
-                style={{ pointerEvents: 'none' }}
-              >
-                {kw.keyword}
-              </text>
-            </motion.g>
-          );
-        })}
-
-        {negativeKeywords.map((kw, i) => {
-          const { x, y } = negPos[i];
-          const pw = pillWidth(kw.keyword);
-          const isSelected = selected?.keyword === kw.keyword;
-          const isDimmed = selected !== null && !isSelected;
-          const hasArticles = kw.articles.length > 0;
-          const fill = hasArticles ? NEG_FILL : EMPTY_FILL;
-          const stroke = hasArticles ? NEG_STROKE : EMPTY_STROKE;
-          const textFill = hasArticles ? NEG_TEXT : EMPTY_TEXT;
-
-          return (
-            <motion.g
-              key={kw.keyword}
-              role="button"
-              tabIndex={isDimmed ? -1 : 0}
-              animate={{
-                x: isSelected ? SVG_WIDTH / 2 - x : 0,
-                y: isSelected ? SVG_CENTER_Y - y : 0,
-                opacity: isDimmed ? 0 : 1,
-              }}
-              transition={{ duration: 0.35, ease: 'easeOut' }}
-              whileHover={
-                selected === null && hasArticles ? { scale: 1.08 } : {}
-              }
-              whileFocus={
-                selected === null && hasArticles ? { scale: 1.08 } : {}
-              }
-              style={{
-                transformOrigin: `${x}px ${y}px`,
-                cursor:
-                  selected === null && hasArticles ? 'pointer' : 'default',
-                outline: 'none',
-              }}
-              onClick={(e) => {
-                if (selected === null && hasArticles) {
-                  e.stopPropagation();
-                  setSelected(kw);
-                }
-              }}
-              onKeyDown={(e) => {
-                if (
-                  selected === null &&
-                  hasArticles &&
-                  (e.key === 'Enter' || e.key === ' ')
-                ) {
-                  e.preventDefault();
-                  e.stopPropagation();
-                  setSelected(kw);
-                }
-              }}
+            </filter>
+            <filter
+              id={negShadowId}
+              x="-30%"
+              y="-30%"
+              width="160%"
+              height="160%"
             >
-              <rect
-                x={x - pw / 2}
-                y={y - PILL_HEIGHT / 2}
-                width={pw}
-                height={PILL_HEIGHT}
-                rx={PILL_RADIUS}
-                fill={fill}
-                stroke={stroke}
-                strokeWidth={1}
-                filter={`url(#${negShadowId})`}
+              <feDropShadow
+                dx="2"
+                dy="2"
+                stdDeviation="2.5"
+                floodColor="rgba(169,204,253,0.35)"
               />
-              <text
-                x={x}
-                y={y}
-                textAnchor="middle"
-                dominantBaseline="middle"
-                fontSize={13}
-                fontWeight="600"
-                fill={textFill}
-                style={{ pointerEvents: 'none' }}
-              >
-                {kw.keyword}
-              </text>
-            </motion.g>
-          );
-        })}
+            </filter>
 
-        <AnimatePresence>
-          {selected !== null &&
-            displayedArticles.map((article, i) => {
-              const { x, y } = newsPos[i];
-              const title = truncateTitle(article.title);
-              const pl = x - NEWS_PILL_WIDTH / 2;
-              const pt = y - NEWS_PILL_HEIGHT / 2;
+            <filter
+              id={newsShadowId}
+              x="-20%"
+              y="-40%"
+              width="140%"
+              height="180%"
+            >
+              <feDropShadow
+                dx="0"
+                dy="3"
+                stdDeviation="5"
+                floodColor="rgba(0,0,0,0.10)"
+              />
+            </filter>
+          </defs>
 
-              return (
-                <motion.g
-                  key={`news-${article.articleId}`}
-                  role="button"
-                  tabIndex={0}
-                  initial={{ opacity: 0, scale: 0.75 }}
-                  animate={{ opacity: 1, scale: 1 }}
-                  exit={{ opacity: 0, scale: 0.75 }}
-                  transition={{
-                    duration: 0.22,
-                    delay: 0.12 + i * 0.05,
-                    ease: 'easeOut',
-                  }}
-                  whileHover={{ scale: 1.04 }}
-                  whileFocus={{ scale: 1.04 }}
-                  style={{
-                    cursor: 'pointer',
-                    transformOrigin: `${x}px ${y}px`,
-                  }}
-                  onClick={(e) => {
+          <AnimatePresence>
+            {selected !== null && (
+              <motion.rect
+                key="dim"
+                x={0}
+                y={0}
+                width={SVG_WIDTH}
+                height={SVG_HEIGHT}
+                initial={{ opacity: 0 }}
+                animate={{ opacity: 1 }}
+                exit={{ opacity: 0 }}
+                transition={{ duration: 0.3 }}
+                fill={
+                  isPos ? 'rgba(255,245,247,0.94)' : 'rgba(245,250,255,0.94)'
+                }
+                style={{ cursor: 'default' }}
+                onClick={(e) => {
+                  e.stopPropagation();
+                  close();
+                }}
+              />
+            )}
+          </AnimatePresence>
+
+          {positiveKeywords.map((kw, i) => {
+            const { x, y } = posPos[i];
+            const pw = pillWidth(kw.keyword);
+            const isSelected = selected?.keyword === kw.keyword;
+            const isDimmed = selected !== null && !isSelected;
+            const hasArticles = kw.articles.length > 0;
+            const fill = hasArticles ? POS_FILL : EMPTY_FILL;
+            const stroke = hasArticles ? POS_STROKE : EMPTY_STROKE;
+            const textFill = hasArticles ? POS_TEXT : EMPTY_TEXT;
+
+            return (
+              <motion.g
+                key={kw.keyword}
+                role="button"
+                tabIndex={isDimmed ? -1 : 0}
+                animate={{
+                  x: isSelected ? SVG_WIDTH / 2 - x : 0,
+                  y: isSelected ? SVG_CENTER_Y - y : 0,
+                  opacity: isDimmed ? 0 : 1,
+                }}
+                transition={{ duration: 0.35, ease: 'easeOut' }}
+                whileHover={
+                  selected === null && hasArticles ? { scale: 1.08 } : {}
+                }
+                whileFocus={
+                  selected === null && hasArticles ? { scale: 1.08 } : {}
+                }
+                style={{
+                  transformOrigin: `${x}px ${y}px`,
+                  cursor:
+                    selected === null && hasArticles ? 'pointer' : 'default',
+                  outline: 'none',
+                }}
+                onClick={(e) => {
+                  if (selected === null && hasArticles) {
                     e.stopPropagation();
-                    onNewsClick(article.articleId);
-                  }}
-                  onKeyDown={(e) => {
-                    if (e.key === 'Enter' || e.key === ' ') {
-                      e.preventDefault();
+                    setSelected(kw);
+                  }
+                }}
+                onKeyDown={(e) => {
+                  if (
+                    selected === null &&
+                    hasArticles &&
+                    (e.key === 'Enter' || e.key === ' ')
+                  ) {
+                    e.preventDefault();
+                    e.stopPropagation();
+                    setSelected(kw);
+                  }
+                }}
+              >
+                <rect
+                  x={x - pw / 2}
+                  y={y - PILL_HEIGHT / 2}
+                  width={pw}
+                  height={PILL_HEIGHT}
+                  rx={PILL_RADIUS}
+                  fill={fill}
+                  stroke={stroke}
+                  strokeWidth={1}
+                  filter={`url(#${posShadowId})`}
+                />
+                <text
+                  x={x}
+                  y={y}
+                  textAnchor="middle"
+                  dominantBaseline="middle"
+                  fontSize={13}
+                  fontWeight="600"
+                  fill={textFill}
+                  style={{ pointerEvents: 'none' }}
+                >
+                  {kw.keyword}
+                </text>
+              </motion.g>
+            );
+          })}
+
+          {negativeKeywords.map((kw, i) => {
+            const { x, y } = negPos[i];
+            const pw = pillWidth(kw.keyword);
+            const isSelected = selected?.keyword === kw.keyword;
+            const isDimmed = selected !== null && !isSelected;
+            const hasArticles = kw.articles.length > 0;
+            const fill = hasArticles ? NEG_FILL : EMPTY_FILL;
+            const stroke = hasArticles ? NEG_STROKE : EMPTY_STROKE;
+            const textFill = hasArticles ? NEG_TEXT : EMPTY_TEXT;
+
+            return (
+              <motion.g
+                key={kw.keyword}
+                role="button"
+                tabIndex={isDimmed ? -1 : 0}
+                animate={{
+                  x: isSelected ? SVG_WIDTH / 2 - x : 0,
+                  y: isSelected ? SVG_CENTER_Y - y : 0,
+                  opacity: isDimmed ? 0 : 1,
+                }}
+                transition={{ duration: 0.35, ease: 'easeOut' }}
+                whileHover={
+                  selected === null && hasArticles ? { scale: 1.08 } : {}
+                }
+                whileFocus={
+                  selected === null && hasArticles ? { scale: 1.08 } : {}
+                }
+                style={{
+                  transformOrigin: `${x}px ${y}px`,
+                  cursor:
+                    selected === null && hasArticles ? 'pointer' : 'default',
+                  outline: 'none',
+                }}
+                onClick={(e) => {
+                  if (selected === null && hasArticles) {
+                    e.stopPropagation();
+                    setSelected(kw);
+                  }
+                }}
+                onKeyDown={(e) => {
+                  if (
+                    selected === null &&
+                    hasArticles &&
+                    (e.key === 'Enter' || e.key === ' ')
+                  ) {
+                    e.preventDefault();
+                    e.stopPropagation();
+                    setSelected(kw);
+                  }
+                }}
+              >
+                <rect
+                  x={x - pw / 2}
+                  y={y - PILL_HEIGHT / 2}
+                  width={pw}
+                  height={PILL_HEIGHT}
+                  rx={PILL_RADIUS}
+                  fill={fill}
+                  stroke={stroke}
+                  strokeWidth={1}
+                  filter={`url(#${negShadowId})`}
+                />
+                <text
+                  x={x}
+                  y={y}
+                  textAnchor="middle"
+                  dominantBaseline="middle"
+                  fontSize={13}
+                  fontWeight="600"
+                  fill={textFill}
+                  style={{ pointerEvents: 'none' }}
+                >
+                  {kw.keyword}
+                </text>
+              </motion.g>
+            );
+          })}
+
+          <AnimatePresence>
+            {selected !== null &&
+              displayedArticles.map((article, i) => {
+                const { x, y } = newsPos[i];
+                const title = truncateTitle(article.title);
+                const pl = x - NEWS_PILL_WIDTH / 2;
+                const pt = y - NEWS_PILL_HEIGHT / 2;
+
+                return (
+                  <motion.g
+                    key={`news-${article.articleId}`}
+                    role="button"
+                    tabIndex={0}
+                    initial={{ opacity: 0, scale: 0.75 }}
+                    animate={{ opacity: 1, scale: 1 }}
+                    exit={{ opacity: 0, scale: 0.75 }}
+                    transition={{
+                      duration: 0.22,
+                      delay: 0.12 + i * 0.05,
+                      ease: 'easeOut',
+                    }}
+                    whileHover={{ scale: 1.04 }}
+                    whileFocus={{ scale: 1.04 }}
+                    style={{
+                      cursor: 'pointer',
+                      transformOrigin: `${x}px ${y}px`,
+                    }}
+                    onClick={(e) => {
                       e.stopPropagation();
                       onNewsClick(article.articleId);
-                    }
-                  }}
-                >
-                  <rect
-                    x={pl}
-                    y={pt}
-                    width={NEWS_PILL_WIDTH}
-                    height={NEWS_PILL_HEIGHT}
-                    rx={NEWS_PILL_RADIUS}
-                    fill="white"
-                    stroke={drillStroke}
-                    strokeWidth={1}
-                    filter={`url(#${newsShadowId})`}
-                  />
-                  <svg
-                    x={pl + 8}
-                    y={y - NEWS_ICON_SIZE / 2}
-                    width={NEWS_ICON_SIZE}
-                    height={NEWS_ICON_SIZE}
-                    style={{ pointerEvents: 'none', overflow: 'visible' }}
+                    }}
+                    onKeyDown={(e) => {
+                      if (e.key === 'Enter' || e.key === ' ') {
+                        e.preventDefault();
+                        e.stopPropagation();
+                        onNewsClick(article.articleId);
+                      }
+                    }}
                   >
-                    <IoDocumentTextOutline
+                    <rect
+                      x={pl}
+                      y={pt}
+                      width={NEWS_PILL_WIDTH}
+                      height={NEWS_PILL_HEIGHT}
+                      rx={NEWS_PILL_RADIUS}
+                      fill="white"
+                      stroke={drillStroke}
+                      strokeWidth={1}
+                      filter={`url(#${newsShadowId})`}
+                    />
+                    <svg
+                      x={pl + 8}
+                      y={y - NEWS_ICON_SIZE / 2}
                       width={NEWS_ICON_SIZE}
                       height={NEWS_ICON_SIZE}
-                      color={drillAccent}
-                    />
-                  </svg>
-                  <text
-                    x={pl + 8 + NEWS_ICON_SIZE + 5}
-                    y={y + 0.5}
-                    dominantBaseline="middle"
-                    fontSize={11}
-                    fontWeight="500"
-                    fill="#374151"
-                    style={{ pointerEvents: 'none' }}
-                  >
-                    {title}
-                  </text>
-                </motion.g>
-              );
-            })}
-        </AnimatePresence>
-      </Svg>
-    </Container>
+                      style={{ pointerEvents: 'none', overflow: 'visible' }}
+                    >
+                      <IoDocumentTextOutline
+                        width={NEWS_ICON_SIZE}
+                        height={NEWS_ICON_SIZE}
+                        color={drillAccent}
+                      />
+                    </svg>
+                    <text
+                      x={pl + 8 + NEWS_ICON_SIZE + 5}
+                      y={y + 0.5}
+                      dominantBaseline="middle"
+                      fontSize={11}
+                      fontWeight="500"
+                      fill="#374151"
+                      style={{ pointerEvents: 'none' }}
+                    >
+                      {title}
+                    </text>
+                  </motion.g>
+                );
+              })}
+          </AnimatePresence>
+        </Svg>
+      </Container>
+    </>
   );
 }
